# server/api/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Smart Sign Caputer API",
        "version": "1.0.0"
    },
    "host": "ssc-dev.smarthealth.at",
    "schemes": [
        "https"
    ],
    "basePath": "/api",
    "produces": [
        "application/json"
    ],
    "paths": {
...
        "/users/{id}": {
            "get": {
                "summary": "User",
                "description": "Users registered at the app\n",
                "tags": [
                    "Users"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the User to fetch",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Specific Users",
                        "schema": {
                            "type": "object",
                            "items": {
                                "$ref": "#/definitions/User"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "User": {
            "properties": {
                "id": {
                    "type": "integer",
                    "description": "Unique identifier representing a specific user."
                },
                "user_name": {
                    "type": "string",
                    "description": "User name of the user."
                },
                "email": {
                    "type": "string",
                    "description": "Email address of the user."
                },
                "first_name": {
                    "type": "string",
                    "description": "First name of the user."
                },
                "last_name": {
                    "type": "string",
                    "description": "Last name of the user."
                }
            }
        },
...
    }
}